fix for IBM S/390
git-svn-id: svn+ssh://mail.hekkelman.net/home/maarten/svnroot/libzeep/trunk@445 31ceaa29-a5dd-4727-bcb8-6a3098495666

+++ src/webapp-el.cpp
@@ -1547,7 +1547,7 @@ object interpreter::parse_primary_expr()
 				{
 					match(m_lookahead);
 					if (result.type() == object::array_type and (m_token_string == "count" or m_token_string == "length"))
-						result = object(result.count());
+						result = object((uint32)result.count());
 					else
 						result = const_cast<const object&>(result)[m_token_string];
 					match(elt_object);
